#72 Fix form save
