+++ app/src/main/ui/package.json
@@ -25,7 +25,7 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@cloudogu/ces-theme-tailwind": "^0.0.17",
+    "@cloudogu/ces-theme-tailwind": "^0.0.21",
     "@mertasan/tailwindcss-variables": "^2.5.1",
     "@types/node": "^18.11.18",
     "@types/react": "^18.0.26",
